optimize the code
